requirements/common.txt
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-compressor==2.2
django-filer==1.3.2
django-filter==1.1.0
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-mptt==0.8.7
django-parler==1.8
django-polymorphic==1.3
django-post-office==3.0.3
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-treebeard==4.2.0
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-file==2.2.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.4.0
djangorestframework==3.7.7
django-formtools==2.1
drf-haystack==1.6.1
certifi==2017.11.5
chardet==3.0.4
easy-thumbnails==2.5
elasticsearch==1.7.0
hiredis==0.2.0
html5lib==0.9999999
jsonfield==1.0.3
libsass==0.12.3
oauthlib==1.0.3
Pillow==4.3.0
idna==2.6
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2015.7
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
six==1.11.0
Unidecode==0.4.18
urllib3==1.22
virtualenv==13.1.2
wheel==0.24.0
bs4==0.0.1
